fix warnings
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@1049 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dope/dope_main.ss/src/dope_main/OdbcPersistor.cpp
@@ -211,7 +211,7 @@ void OdbcPersistor::Store(const Safir::Dob::Typesystem::EntityId entityId,
             {
                 if (small)
                 {
-                    m_storeBinarySmallDataParam.SetValue(&bin[0],bin.size());
+                    m_storeBinarySmallDataParam.SetValue(&bin[0],static_cast<unsigned int>(bin.size()));
                     m_storeBinaryDataParam.SetNull();
                 }
                 else
